(task_database) Use size_t for size to prevent narrowing issue

+++ tasks/task_database.c
@@ -251,7 +251,7 @@ static int intfstream_get_serial(intfstream_t *fd, char *serial, size_t serial_l
 }
 
 static bool intfstream_file_get_serial(const char *name,
-      uint64_t offset, uint64_t size, char *serial, size_t serial_len)
+      uint64_t offset, size_t size, char *serial, size_t serial_len)
 {
    int rv;
    uint8_t *data     = NULL;
@@ -273,12 +273,12 @@ static bool intfstream_file_get_serial(const char *name,
    if (file_size < 0)
       goto error;
 
-   if (offset != 0 || size < (uint64_t) file_size)
+   if (offset != 0 || size < (size_t) file_size)
    {
       if (intfstream_seek(fd, (int64_t)offset, SEEK_SET) == -1)
          goto error;
 
-      data = (uint8_t*)malloc((size_t)size);
+      data = (uint8_t*)malloc(size);
 
       if (intfstream_read(fd, data, size) != (int64_t) size)
       {
@@ -312,10 +312,10 @@ error:
 static int task_database_cue_get_serial(const char *name, char* serial, size_t serial_len)
 {
    char track_path[PATH_MAX_LENGTH];
-   uint64_t offset                  = 0;
-   uint64_t size                    = 0;
+   uint64_t offset  = 0;
+   size_t size      = 0;
 
-   track_path[0]                    = '\0';
+   track_path[0]    = '\0';
 
    if (cue_find_track(name, true, &offset, &size, track_path,
             sizeof(track_path)) < 0)
@@ -429,7 +429,7 @@ static int task_database_cue_get_crc(const char *name, uint32_t *crc)
 {
    char track_path[PATH_MAX_LENGTH];
    uint64_t offset  = 0;
-   uint64_t size    = 0;
+   size_t size      = 0;
 
    track_path[0]    = '\0';
 
@@ -443,7 +443,7 @@ static int task_database_cue_get_crc(const char *name, uint32_t *crc)
       return 0;
    }
 
-   return intfstream_file_get_crc(track_path, offset, (size_t)size, crc);
+   return intfstream_file_get_crc(track_path, offset, size, crc);
 }
 
 static int task_database_gdi_get_crc(const char *name, uint32_t *crc)
